portia_demo/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
portia-sdk-python==0.2.6
openai==1.6.1
anthropic==0.7.4
gunicorn==21.2.0
pytest==7.4.0
requests==2.31.0
typing-extensions==4.9.0 